requirements.txt
opencv-python
google-generativeai
requests
python-dotenv
